Remove callhome core source code
