+++ pom.xml
@@ -14,7 +14,8 @@
  See the License for the specific language governing permissions and
  limitations under the License.
 -->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
         <groupId>org.wso2</groupId>
         <artifactId>wso2</artifactId>
@@ -24,7 +25,7 @@
     <groupId>org.wso2.carbon.callhome</groupId>
     <artifactId>callhome</artifactId>
     <name>Call Home Client</name>
-    <version>4.5.x_1.0.4-SNAPSHOT</version>
+    <version>4.4.x_1.0.5-SNAPSHOT</version>
     <packaging>bundle</packaging>
     <modelVersion>4.0.0</modelVersion>
 
@@ -36,25 +37,13 @@
     </scm>
 
     <properties>
-        <org.apache.httpcomponents.httpclient>4.5.11</org.apache.httpcomponents.httpclient>
-        <org.apache.httpcomponents.httpcore>4.4.13</org.apache.httpcomponents.httpcore>
         <org.eclipse.platform>3.14.0</org.eclipse.platform>
         <org.eclipse.osgi.services>3.3.100.v20130513-1956</org.eclipse.osgi.services>
         <org.wso2.carbon.core>4.5.2</org.wso2.carbon.core>
-        <org.yaml.snakeyaml>1.23</org.yaml.snakeyaml>
+        <org.wso2.carbon.callhome.core>1.0.3</org.wso2.carbon.callhome.core>
     </properties>
 
     <dependencies>
-        <dependency>
-            <groupId>org.apache.httpcomponents</groupId>
-            <artifactId>httpclient</artifactId>
-            <version>${org.apache.httpcomponents.httpclient}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.httpcomponents</groupId>
-            <artifactId>httpcore</artifactId>
-            <version>${org.apache.httpcomponents.httpcore}</version>
-        </dependency>
         <dependency>
             <groupId>org.eclipse.platform</groupId>
             <artifactId>org.eclipse.osgi</artifactId>
@@ -71,11 +60,12 @@
             <version>${org.wso2.carbon.core}</version>
         </dependency>
         <dependency>
-            <groupId>org.yaml</groupId>
-            <artifactId>snakeyaml</artifactId>
-            <version>${org.yaml.snakeyaml}</version>
+            <groupId>org.wso2.carbon.callhome</groupId>
+            <artifactId>core</artifactId>
+            <version>${org.wso2.carbon.callhome.core}</version>
         </dependency>
     </dependencies>
+
     <build>
         <plugins>
             <plugin>
@@ -109,12 +99,12 @@
                     <instructions>
                         <Bundle-Vendor>WSO2 Inc</Bundle-Vendor>
                         <Bundle-SymbolicName>${project.artifactId}</Bundle-SymbolicName>
-                        <Bundle-Activator>org.wso2.callhome.internal.CallHomeActivator</Bundle-Activator>
+                        <Bundle-Activator>org.wso2.carbon.callhome.internal.CallHomeActivator</Bundle-Activator>
                         <Private-Package>
-                            org.wso2.callhome.*
+                            org.wso2.carbon.callhome.*
                         </Private-Package>
                         <Export-Package>
-                            !org.wso2.callhome.*
+                            !org.wso2.carbon.callhome.*
                         </Export-Package>
                         <Import-Package>
                             org.wso2.carbon.base.api; version="[1.0.0,2.0.0)",
@@ -125,7 +115,7 @@
                             org.osgi.service.component; version="[1.2.0,2.0.0)",
                         </Import-Package>
                         <Embed-Dependency>
-                            snakeyaml,httpcore,httpclient;scope=compile|runtime;inline=false
+                            core;scope=compile|runtime;inline=false
                         </Embed-Dependency>
                     </instructions>
                 </configuration>
